requirements.txt
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
mailchimp-marketing==3.0.80
rapidfuzz>=3.0.0
